wl12xx: fw 3.2 - change template max/default size
The new fw changed the max template size.
However, we should use the max size only when needed, as it
consumes some of the chip's memory.

Thus, by default initialize the templates to the default size.
Initialize to the maximum size only when required.

NOTE: it looks like we have some sdio(?) problem that prevents
us from sending a command with the max template size.