wl12xx: Use dev_hlid for auth and assoc req (EXPERIMENTAL)
On roaming, the auth and assoc req are sent with the sta
hlid. this is wrong, as the sta wasn't started yet with
the new ap params. however, it shoudln't really matter as
the sta was already tuned to the right channel.
However, we have some roaming issues in some special cases,
so check whether using the dev hlid solved them.