wl12xx: add existing stations after starting ap
In the normal ap flow, the ap is started and later on
the stations are added. however, when the vif if reconfigured,
the stations are configured before the ap is started (no
beaconing). the fw doesn't handle well this situation.

Solve it by iterating through the stations right after
starting the ap, and adding any not-already-added sta.

Signed-off-by: Eliad Peller <eliad@wizery.com>