wl18xx: 11b rate issue fix
when working as STA, the rate policy for the auth/assoc frames
includes only the default rates, and does not refer to the AP's basic ones.
the solution is that mac80211 informs the driver about bss basic rate change
before authentication is started (taken only the min basic rate) and configures its policy accordingly.

Signed-off-by: Lior Cohen <liorc@ti.com>