wl12xx: add chip_family to platform data struct
To abstract the bus modules from the actual driver that handles the
chip, we can pass the chip_family struct in the platform data.  This
string is used by the bus modules to name the platform device they
create.  The same string is used by the platform driver to get probed.

If not specified, we use "wl12xx" as default.

Signed-off-by: Luciano Coelho <coelho@ti.com>