wlcore: add module skeleton
Just add the module main.c file with a skeleton for init and friends.

This module will be used to hold the common code for TI's WLAN
drivers.  It will be the part that can be reused across differnt chip
families.

This patch also creates a subdirectory that will hold all TI WLAN
modules, called drivers/net/wireless/ti.

Signed-off-by: Luciano Coelho <coelho@ti.com>