18xx: use sg dma for Tx in 18xx cards
Add a quirk to enable sg dma for the Tx path in 18xx cards. Initialize
a sg-list and add buffers there instead of the aggregation buffer. When
flushing the data use the new sg_write API call.

TODO: remove manual 4-byte alignment of skb base if/when FW support is
there for it.