wlcore: add implementation to send commands to the firmware
Add command definitions and a function to send commands to the
firmware.

To support the send command function, a couple of registers (used to
trigger the commands) has been added to the register table.

The bits used as triggers also vary from hardware to hardware, so we
add a trigger table where the hardware-specific trigger values will be
added.

Signed-off-by: Luciano Coelho <coelho@ti.com>