wlcore: add chip_family to plat data and magic to ensure the correct struct
Add a signature magic number in place of the irq in the
wl12xx_platform_data struct to make sure we have the correct version
of the structure (defined in the board files).

Add chip_family to the structure and use in the bus modules in order
to name the platform device created and map it to the appopriate
driver to be loaded.

Signed-off-by: Luciano Coelho <coelho@ti.com>