wlcore: merge FW status structure from both chip families
Create a union FW status to fit all chip families. Modify functions that
access parts of the status to operate according to chip family.

Note: we don't include hw ops for accessing parts of the FW, as there's
a chance for changes in the 18xx side that will align the two firmwares.

Signed-off-by: Arik Nemtsov <arik@wizery.com>