wl18xx: WIP adding register table
Add the register table with the appropriate values in wl18xx.  Some of
the registers are commented out and may not be necessary in wl18xx, so
they may have to be moved out of wlcore and into wl12xx.

This is still work in progress and needs to be amended.

Signed-off-by: Luciano Coelho <coelho@ti.com>