wlcore/wl12xx: move top initialization to wl12xx
The top registers initialization is very specific to the actual
hardware used, even the way in which we read from and write to the top
registers varies from chip to chip.  This patch moves all top
registers initialization to wl12xx.  Also add a few ops for the wlcore
module to call at the right time.

Signed-off-by: Luciano Coelho <coelho@ti.com>