wl18xx: change Tx-complete indication in FW status for 18xx cards
Currently the number of freed Tx descriptors is given in an index in the
FW status for 18xx cards. During fast Tx, all Tx descriptors can be
freed in a single interrupt. With the previous FW-status this would have
caused the driver to believe no descriptors were freed because of the
wrap-around.

Add a dummy Tx-desc-freed entry to the FW-status so we avoid wrap-around
for this case.

Signed-off-by: Arik Nemtsov <arik@wizery.com>