wl12xx: unset TX alignment for wl127x chips and don't set for wl128x
The call to wl1271_set_block_size() decides whether SDIO block size
alignment can be used or not.  In case we're using SPI, we can't use
the block size alignment, so the function returns false.  The wl12xx
lower driver was setting the quirk for wl128x and not touching the bit
for wl127x.  Instead, we should unset the bit for wl127x and not touch
it for wl128x.

Signed-off-by: Luciano Coelho <coelho@ti.com>