wl18xx: don't set the TX blocksize alignment quirk in the lower driver
The call to wl1271_set_block_size() decides whether SDIO block size
alignment can be used or not.  In case we're using SPI, we can't use
the block size alignment, so the function returns false.  The wl18xx
lower driver was setting the quirk, but we shouldn't touch this bit
here.

Signed-off-by: Luciano Coelho <coelho@ti.com>