wl12xx: fix wl12xx_sdio_interrupt() deadlock
before wl12xx_sdio_interrupt() calls wl1271_irq() directly,
it releases the host, and then reclaim it again. This might
deadlock on wl12xx_free_sdio_irq(), as it both claims the
host and waits for wl12xx_sdio_interrupt() to finish (which
tries to claim the already-claimed host).

Solve it by defering the wl1271_irq() call to a different
work. however, now we have to explicitly disable the irq
until wl1271_irq() is called, so add some code to handle
it...

Signed-off-by: Eliad Peller <eliad@wizery.com>