wlcore/wl12xx: move part of the exported elements to a separate struct
To help reorganizing what is exported to the lower drivers, we create
a new struct that contains the exported part.  For now, to simplify
the commit and reduce the complexity of the patch, we just add the new
struct at the top of the existing wl1271 struct and change all the
files that access it so that they now access the elements in the
sub-structure.

Signed-off-by: Luciano Coelho <coelho@ti.com>