wlcore/wl12xx: change read_data op to prepare_read
The only difference in the read_data operations is that some chips
need to prepare the data to be read before reading.  So instead of
having a mandatory read_data operation, we now have an option
prepare_data operation that only needs to be implemented for chips
that require it.

In the wl12xx lower driver, we only set the prepare_data operation for
wl127x chips.

With this change, we don't need to export the aggr_buf element and it
also avoid unnecessary operation calls.

Signed-off-by: Luciano Coelho <coelho@ti.com>