mac80211: API to enable send ADDTS request (INTERNAL)
When using WMM, the AP can set an access category queue with the
admission control mandatroy bit set to 1, in this case
in order to send traffic on such an acces category queue
the device will need to send an ADDTS with a tspec IE in order
to admit the queue.

[Merged, moved new wmm elements from local to sdata, moved debugfs
entry to debugfs_netdev and a few other related things. -- Luca]

Signed-off-by: Yoni Divinsky <yoni.divinsky@ti.com>
Signed-off-by: Luciano Coelho <coelho@ti.com>