wl12xx: Register lock-less IBI handlers *EXPERIMENTAL*
wl->mutex is normally acquired before claiming the mmc host. In the case of IBI, the handler is called with
the host already claimed. This means that acquiring the mutex in the handler will result in an incorrect
locking order, which could lead to a deadlock.

Fix this by registering the handler as a lock-less one, and claim the host after acquiring the mutex.

Signed-off-by: Ido Reis <idor@ti.com>
Signed-off-by: Eliad Peller <eliad@wizery.com>