wlcore/wl12xx: add private are in fw_status (MOVE DOWN)
Most of the fw status structure is the same across chips, but each
chip can have its own specific data to pass up.  This commit adds a
private area in the fw status that can be used by the lower drivers.
Also add an exported parameter to tell how large the private area
should be (in wl12xx currently it's zero).

Signed-off-by: Luciano Coelho <coelho@ti.com>