wlcore/wl12xx/wl18xx: clean PG_VER and MAC address reading (SQUASH/SPLIT)
This patch cleans up the PG_VER reading for wl12xx and wl18xx (add a
new op for it) and the MAC from fuse reading for wl12xx.

This got broken when rebasing the wl18xx work on top of R5_3.2.

This patch should be squashed back to the point when it got broken and
fixed there directly.  We also need to split the wl18xx part out of
it.

Signed-off-by: Luciano Coelho <coelho@ti.com>