wlcore: add fw_name module parameters for overriding
In order to support multiple concurrent firmware projects, we need to
allow the used firmware names to be changed at runtime.  This patch
adds module parameters for the FW names and checks if the name is the
same, in which case, we don't need to perform a fw change (with HW
recovery).

This is temporary and should be cleaned up for upstream.  The upstream
code with fw_type and such is already a bit different, so we need to
rebase and re-do this properly.  There are other, probably cleaner,
ways to do this.

Signed-off-by: Luciano Coelho <coelho@ti.com>