wlcore/wl12xx/wl18xx: move version prints to module init
It's good to print out the driver version as early as possible, so we
can also catch problems with the probe.  Move the version info to init.

Also, change "compilation time" to "timestamp", because it doesn't
mean the compile time anymore but the modification time of the source
file that generates it.

Signed-off-by: Luciano Coelho <coelho@ti.com>