wl18xx: use a single dimension for tx_agg_vs_rate in fw_stats (SQUASH)
This fixes commit 6c5ecb12 (wl18xx: update fw statistics) where we
were printing values from out-of-bounds memory areas.

Flatten the array into a single dimension to make it possible to print
it out using the existing macro.  The only drawback is that [1][0]
will be printed as [16] instead, but this is better than printing
random pointers.

Signed-off-by: Luciano Coelho <coelho@ti.com>