wlcore/wl12xx/wl18xx: support setting MAC addresses in conf
Allow the MAC addresses for a device to be set in conf. These addresses
default to 0 and can be overridden by the NVS file or FUSE.

Add a num_macs element per HW driver, to account for a different number
of MAC addresses reserved for different platforms (12xx vs 18xx).

bump up WLCORE_CONF_VERSION to reflect the changes in conf struct.

Signed-off-by: Arik Nemtsov <arik@wizery.com>