mac80211_hwsim: use channel contexts
To use mac80211_hwsim for testing channel contexts it
has to support them, and for that it has to support
hw scan and hw-remain-on-channel.

Since it's pure software, the off-channel activities
are really not off-channel but listening and sending
on a second channel. Also, the multi-channel isn't
really doing TDM, it's just on both channels at the
same time.

Note that this breaks wmediumd right now as it doesn't
propagate the channel.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>